src/backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf2==3.0.1
langchain==0.0.200
boto3==1.28.38
redis==4.5.5
celery==5.3.0
jinja2==3.1.2
weasyprint==59.0
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.1